+++ examples/simple.js
@@ -2,6 +2,6 @@
 
 const awsParamEnv = require( '..' /*'aws-param-env'*/ );
 
-awsParamEnv.load( '/services/storage-system/production/env' );
+awsParamEnv.load( '/services/storage-system/production/env', { region: 'us-east-1' } );
 
 console.log( process.env );