+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Change Log
 
+## 1.2.0 (2018-01-08)
+
+Improved:
+
+* Added options to define region
+
 ## 1.1.0 (2017-09-18)
 
 New:
 
-* `load()` function that executes the query with a given path 
+* `load()` function that executes the query with a given path
 
 ## 1.0.0 (2017-09-13)